+++ CHANGES.rst
@@ -38,6 +38,9 @@ API Changes
 Cubeviz
 ^^^^^^^
 
+- New tool for visualizing spectrum at a pixel's coordinate location
+  in the image viewer [#1317]
+
 Imviz
 ^^^^^